Extract all logic to function
